Enabled separate nav graph for main screens
