+++ app/src/main/java/in/androidplay/trackme/ui/compose/components/BottomNavigationBar.kt
@@ -74,7 +74,7 @@ fun BottomNavigationBar(
                                 contentDescription = stringResource(id = screen.label)
                             )
                     },
-                    alwaysShowLabel = true,
+                    alwaysShowLabel = false,
                     selected = isSelected,
                     label = { Text(text = screen.route) },
                     colors = NavigationBarItemDefaults.colors(
