+++ app/src/main/java/in/androidplay/trackme/ui/navigation/AppNavigation.kt
@@ -2,12 +2,8 @@ package `in`.androidplay.trackme.ui.navigation
 
 import androidx.compose.runtime.Composable
 import androidx.navigation.compose.NavHost
-import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import `in`.androidplay.trackme.ui.compose.components.sharedViewModel
-import `in`.androidplay.trackme.ui.compose.home.HomeScreen
-import `in`.androidplay.trackme.ui.compose.home.HomeViewModel
-import `in`.androidplay.trackme.util.Constants.HOME_GRAPH
+import `in`.androidplay.trackme.util.Constants.MAIN_GRAPH
 import `in`.androidplay.trackme.util.Constants.ONBOARDING_GRAPH
 
 @Composable
@@ -15,11 +11,9 @@ fun AppNavigation(isFirstAppOpen: Boolean) {
     val navController = rememberNavController()
     NavHost(
         navController = navController,
-        startDestination = if (isFirstAppOpen) ONBOARDING_GRAPH else HOME_GRAPH
+        startDestination = if (isFirstAppOpen) ONBOARDING_GRAPH else MAIN_GRAPH
     ) {
         onboardingNavGraph(navController = navController)
-        composable(route = HOME_GRAPH) {
-            HomeScreen()
-        }
+        mainNavGraph()
     }
 }
