+++ app/src/main/java/in/androidplay/trackme/ui/navigation/HomeNavGraph.kt
@@ -5,12 +5,11 @@ import androidx.compose.ui.Modifier
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
-import `in`.androidplay.trackme.ui.compose.components.sharedViewModel
 import `in`.androidplay.trackme.ui.compose.home.HomeScreenContent
 import `in`.androidplay.trackme.ui.compose.home.HomeViewModel
-import `in`.androidplay.trackme.ui.compose.home.SettingsScreenContent
-import `in`.androidplay.trackme.ui.compose.home.StatisticScreenContent
-import `in`.androidplay.trackme.util.Constants.HOME_GRAPH
+import `in`.androidplay.trackme.ui.compose.settings.SettingsScreenContent
+import `in`.androidplay.trackme.ui.compose.stats.StatisticScreenContent
+import `in`.androidplay.trackme.util.Constants.MAIN_GRAPH
 
 @Composable
 fun HomeNavGraph(
@@ -20,7 +19,7 @@ fun HomeNavGraph(
 ) {
     NavHost(
         navController = navController,
-        route = HOME_GRAPH,
+        route = MAIN_GRAPH,
         startDestination = BottomNavItem.HomeScreen.route
     ) {
         composable(route = BottomNavItem.HomeScreen.route) {
